last_seven/requirements.txt
mysql-connector-python==8.3.0
pandas==2.2.0